requirements.txt
altgraph==0.17.4
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
lxml==5.3.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
numpy==2.1.2
oauth2client==4.1.3
opencv-python==4.10.0.84
packaging==24.1
paramiko==3.5.0
pefile==2023.2.7
pillow==10.4.0
protobuf==5.29.0rc2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyNaCl==1.5.0
pyparsing==3.2.0
Pyrebase4==4.8.0
pytesseract==0.3.13
python-docx==1.1.2
python-jwt==4.1.0
pywin32-ctypes==0.2.3
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
